requirements.txt
google-cloud-storage
google-cloud-speech
google-cloud-aiplatform
google-api-core
functions-framework
pillow
PyPDF2
python-docx
